backend/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.2
psycopg2-binary==2.9.5
python-dotenv==0.21.0
